+++ crates/go-runtime/zkvm_runtime/runtime.go
@@ -1,13 +1,14 @@
 //go:build mipsle
 // +build mipsle
 
-package zkm_runtime
+package zkvm_runtime
+
 import (
-	"encoding/binary"
 	"crypto/sha256"
+	"encoding/binary"
 	"hash"
-	"unsafe"
 	"reflect"
+	"unsafe"
 )
 
 func SyscallWrite(fd int, write_buf []byte, nbytes int) int
@@ -17,16 +18,18 @@ func SyscallCommit(index int, word uint32)
 func SyscallExit(code int)
 
 var PublicValuesHasher hash.Hash = sha256.New()
-const EMBEDDED_RESERVED_INPUT_REGION_SIZE int = 1024 * 1024 * 1024;
-const MAX_MEMORY int = 0x7ff00000;
-var RESERVED_INPUT_PTR int =  MAX_MEMORY - EMBEDDED_RESERVED_INPUT_REGION_SIZE
+
+const EMBEDDED_RESERVED_INPUT_REGION_SIZE int = 1024 * 1024 * 1024
+const MAX_MEMORY int = 0x7ff00000
+
+var RESERVED_INPUT_PTR int = MAX_MEMORY - EMBEDDED_RESERVED_INPUT_REGION_SIZE
 
 func Read[T any]() T {
 	len := SyscallHintLen()
 	var value []byte
 	capacity := (len + 3) / 4 * 4
-	addr := RESERVED_INPUT_PTR;
-	RESERVED_INPUT_PTR += capacity;
+	addr := RESERVED_INPUT_PTR
+	RESERVED_INPUT_PTR += capacity
 	ptr := unsafe.Pointer(uintptr(addr))
 	value = unsafe.Slice((*byte)(ptr), capacity)
 	var result T
@@ -48,7 +51,6 @@ func Commit[T any](value T) {
 	SyscallWrite(13, bytes, length)
 }
 
-
 //go:linkname RuntimeExit zkvm.RuntimeExit
 func RuntimeExit(code int) {
 	hashBytes := PublicValuesHasher.Sum(nil)
@@ -64,5 +66,5 @@ func RuntimeExit(code int) {
 
 func init() {
 	// 显式引用，防止优化
-	 _ = reflect.ValueOf(RuntimeExit)
+	_ = reflect.ValueOf(RuntimeExit)
 }
