+++ examples/simple-go/guest/go.mod
@@ -2,6 +2,6 @@ module simple-go
 
 go 1.22.5
 
-replace github.com/ProjectZKM/Ziren/crates/go-runtime/zkm_runtime => ../../../crates/go-runtime/zkm_runtime
+replace github.com/ProjectZKM/Ziren/crates/go-runtime/zkvm_runtime => ../../../crates/go-runtime/zkvm_runtime
 
-require github.com/ProjectZKM/Ziren/crates/go-runtime/zkm_runtime v0.0.0-20240817102429-2faba0888c02
+require github.com/ProjectZKM/Ziren/crates/go-runtime/zkvm_runtime v0.0.0-20240817102429-2faba0888c02
